backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
opencv-python==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
web3==6.11.3
dnspython==2.4.2
beautifulsoup4==4.12.2
